+++ tests/testthat/test-create_fusen_rsproject.R
@@ -1,0 +1,77 @@
+# Create a new project
+dummypackage <- tempfile(pattern = "dummy")
+dir.create(dummypackage)
+
+raw_rmd_template <- readLines(system.file("dev-template-full.Rmd", package = "fusen"))
+expected_rmd_template_with_fusen_name <- gsub("<my_package_name>", "foosen", raw_rmd_template)
+
+withr::with_dir(dummypackage, {
+ # browser()
+  test_that("Create fusen", {
+    path_foosen <- file.path(dummypackage, "foosen")
+    path_dev_history <- file.path(path_foosen, "dev", "dev_history.Rmd")
+    create_fusen(path_foosen, name = "full", open = FALSE)
+
+    expect_true(dir.exists(path_foosen))
+    expect_true(file.exists(path_dev_history))
+
+    actual_dev_history <- readLines(path_dev_history)
+    expect_identical(
+      actual_dev_history,
+      expected_rmd_template_with_fusen_name
+    )
+  })
+})
+
+# Test overwrite existing dir ----
+withr::with_dir(dummypackage, {
+  test_that("Overwrite fusen dir is possible with message", {
+    path_foosen <- file.path(dummypackage, "foosen")
+
+    # Can not overwrite
+    expect_error(
+      expect_message(
+      create_fusen(path_foosen, name = "full", open = FALSE),
+      "Could not create fusen project"
+      )
+    )
+
+    # Can overwrite with message
+    expect_message(
+      create_fusen(path_foosen, name = "full", open = FALSE, overwrite = TRUE),
+      "Some files may be overwritten"
+    )
+
+    # Test with rstudioapi ----
+    if (requireNamespace("rstudioapi") && rstudioapi::isAvailable()) {
+      expect_true(file.exists(file.path(path_foosen, "foosen.Rproj")))
+    } else {
+      expect_true(file.exists(file.path(path_foosen, ".here")))
+    }
+  })
+})
+
+
+# Create in a subdirectory ----
+dummysubdir <- tempfile(pattern = "subdir/subdir2/dummy")
+test_that("Can create in a subdirectory", {
+  expect_error(create_fusen(dummysubdir, name = "full", open = FALSE), regexp = NA)
+  expect_true(dir.exists(dummysubdir))
+})
+
+# Test gui ----
+dummygui <- tempfile(pattern = "dummygui")
+the_dir <- dirname(dummygui)
+the_path <- basename(dummygui)
+
+withr::with_dir(the_dir, {
+  test_that("Can create in a subdirectory", {
+    expect_error(create_fusen_gui(the_path, name = "teaching"), regexp = NA)
+    expect_true(dir.exists(dummygui))
+  })
+})
+
+unlink(dummypackage, recursive = TRUE)
+unlink(dummysubdir, recursive = TRUE)
+unlink(dummygui, recursive = TRUE)
+