+++ NEWS.md
@@ -1,6 +1,7 @@
 # v0.2.4
 
 * Update vignette and tests
+* `create_fusen()` to create a {fusen} project from command line or with RStudio new project (@ALanguillaume)
 
 # v0.2.3